backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
Pillow==10.4.0
python-decouple==3.8
dj-database-url==2.2.0
google-generativeai==0.7.2
psycopg2-binary>=2.9.9
gunicorn>=22.0.0
whitenoise>=6.7.0
python-dotenv>=1.0.0
requests>=2.31.0
